Code refactored - deleted sockets and model Analysis
