+++ server/src/controllers/onSubmitEstimation.ts
@@ -1,16 +1,25 @@
 import * as socketio from 'socket.io';
 import { rooms } from '@models/Rooms';
 
-const onSubmitEstimation = (io: socketio.Server, socket: socketio.Socket) => (roomId: string) => {
+interface EstimationResult {
+  result: string;
+  roomId: string;
+}
+
+const onSubmitEstimation = (io: socketio.Server, socket: socketio.Socket) => ({ roomId, result }: EstimationResult) => {
   try {
     const room = rooms.getRoom(roomId);
+
+    room.getAdmin(socket.id);
+    room.getTask().reassignFinalResult(Number(result));
 
     room.archiveTask();
     room.clearVotes();
 
-    const message: string = 'The task was saved in history';
+    const message = `Final result: ${result} in the room: ${roomId}`;
 
     io.to(roomId).emit('FEED', message);
+    io.to(roomId).emit('ESTIMATION_SUBMITTED', room.getTask());
   } catch (ex) {
     console.error(ex);
   }
