+++ server/src/controllers/onUserVote.ts
@@ -17,8 +17,9 @@ const onUserVote = (io: socketio.Server, socket: socketio.Socket) => ({ name, va
 
     if (room.hasEveryoneVoted()) {
       room.getTask().setResults(room.getVotes());
+      room.getTask().analyzeResults();
       message = `Everyone in room ${room.id} voted, votes: ${JSON.stringify(room.getVotes())}`;
-      io.to(roomId).emit('CARDS_REVEALED', room.getVotes());
+      io.to(roomId).emit('CARDS_REVEALED', room.getTask());
     } else {
       message = `${name} has voted in the room: ${roomId}`;
       io.to(roomId).emit('USER_VOTED', user);
